requirements/base.txt
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
defusedxml==0.5.0
Django==2.0.03
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.1.0
django-phonenumber-field==2.1.0
django-rest-auth==0.9.3
django-rest-swagger==2.1.2
djangorestframework==3.9.0
djangorestframework-stubs==0.3.0
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
oauthlib==2.1.0
openapi-codec==1.3.2
phonenumbers==8.10.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python3-openid==3.1.0
pytz==2018.7
pytest==4.0.1
pytest-cov==2.6.0
pytest-django==3.4.4
requests==2.21.0
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
djangorestframework-jwt==1.11.0
